fix(deepseek-v4): remove redundant Option wrapping and u64 casts
The DeepSeek-V4 loader and prefill paths wrapped MLA auxiliary weights
(w_uk_t, w_uv, wq_b_rope, w_qk_absorbed, w_uk_block_diag, w_uv_block_diag)
in Some() only to immediately unwrap_or(NULL) at the call site. These
weights are always built (never optional) in the V4 loader, so the Option
layer was dead code.

Similarly, the V4 prefill paths cast pointer offsets and byte counts to u64
via 'as u64' even though DevicePtr::offset takes

+++ crates/spark-model/src/layers/qwen3_attention/prefill/cache_skip_v4.rs
@@ -99,7 +99,7 @@ impl Qwen3AttentionLayer {
 
         // ── 2. Direct KV projection ──
         let k_contiguous = ctx.buffers.ssm_qkvz();
-        let v_contiguous = k_contiguous.offset((num_tokens * (nkv * hd) as usize * bf16) as u64);
+        let v_contiguous = k_contiguous.offset(num_tokens * (nkv * hd) as usize * bf16);
         ops::dense_gemm(
             ctx.gpu,
             self.dense_gemm_k,
@@ -145,10 +145,10 @@ impl Qwen3AttentionLayer {
             let slot_offset = write_start * 8;
             self.write_kv_cache(
                 ctx.gpu,
-                k_contiguous.offset(k_offset as u64),
-                v_contiguous.offset(v_offset as u64),
+                k_contiguous.offset(k_offset),
+                v_contiguous.offset(v_offset),
                 kv_cache,
-                meta.slot.offset(slot_offset as u64),
+                meta.slot.offset(slot_offset),
                 write_count as u32,
                 nkv,
                 hd,
